Update all consumers to use new reader.
