+++ dotnet/src/Connectors/Connectors.Memory.Qdrant/QdrantVectorStoreRecordMapper.cs
@@ -4,7 +4,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using System.Text.Json;
 using Microsoft.SemanticKernel.Data;
 using Qdrant.Client.Grpc;
 
@@ -17,23 +16,8 @@ namespace Microsoft.SemanticKernel.Connectors.Qdrant;
 internal sealed class QdrantVectorStoreRecordMapper<TRecord> : IVectorStoreRecordMapper<TRecord, PointStruct>
     where TRecord : class
 {
-    /// <summary>The public parameterless constructor for the current data model.</summary>
-    private readonly ConstructorInfo _constructorInfo;
-
-    /// <summary>A property info object that points at the key property for the current model, allowing easy reading and writing of this property.</summary>
-    private readonly PropertyInfo _keyPropertyInfo;
-
-    /// <summary>A list of property info objects that point at the data properties in the current model, and allows easy reading and writing of these properties.</summary>
-    private readonly List<PropertyInfo> _dataPropertiesInfo;
-
-    /// <summary>A list of property info objects that point at the vector properties in the current model, and allows easy reading and writing of these properties.</summary>
-    private readonly List<PropertyInfo> _vectorPropertiesInfo;
-
-    /// <summary>A dictionary that maps from a property name to the configured name that should be used when storing it.</summary>
-    private readonly Dictionary<string, string> _storagePropertyNames;
-
-    /// <summary>A dictionary that maps from a property name to the configured name that should be used when serializing it to json.</summary>
-    private readonly Dictionary<string, string> _jsonPropertyNames = new();
+    /// <summary>A helper to access property information for the current data model and record definition.</summary>
+    private readonly VectorStoreRecordPropertyReader _propertyReader;
 
     /// <summary>A value indicating whether the vectors in the store are named, or whether there is just a single unnamed vector per qdrant point.</summary>
     private readonly bool _hasNamedVectors;
@@ -41,51 +25,42 @@ internal sealed class QdrantVectorStoreRecordMapper<TRecord> : IVectorStoreRecor
     /// <summary>
     /// Initializes a new instance of the <see cref="QdrantVectorStoreRecordMapper{TDataModel}"/> class.
     /// </summary>
-    /// <param name="vectorStoreRecordDefinition">The record definition that defines the schema of the record type.</param>
+    /// <param name="propertyReader">A helper to access property information for the current data model and record definition.</param>
     /// <param name="hasNamedVectors">A value indicating whether the vectors in the store are named, or whether there is just a single unnamed vector per qdrant point.</param>
-    /// <param name="storagePropertyNames">A dictionary that maps from a property name to the configured name that should be used when storing it.</param>
     public QdrantVectorStoreRecordMapper(
-        VectorStoreRecordDefinition vectorStoreRecordDefinition,
-        bool hasNamedVectors,
-        Dictionary<string, string> storagePropertyNames)
+        VectorStoreRecordPropertyReader propertyReader,
+        bool hasNamedVectors)
     {
-        Verify.NotNull(vectorStoreRecordDefinition);
-        Verify.NotNull(storagePropertyNames);
+        Verify.NotNull(propertyReader);
 
         // Validate property types.
-        this._constructorInfo = VectorStoreRecordPropertyReader.GetParameterlessConstructor(typeof(TRecord));
-        var propertiesInfo = VectorStoreRecordPropertyReader.FindProperties(typeof(TRecord), vectorStoreRecordDefinition, supportsMultipleVectors: hasNamedVectors);
-        VectorStoreRecordPropertyVerification.VerifyPropertyTypes(propertiesInfo.DataProperties, QdrantVectorStoreRecordFieldMapping.s_supportedDataTypes, "Data", supportEnumerable: true);
-        VectorStoreRecordPropertyVerification.VerifyPropertyTypes(propertiesInfo.VectorProperties, QdrantVectorStoreRecordFieldMapping.s_supportedVectorTypes, "Vector");
+        propertyReader.VerifyHasParameterlessConstructor();
+        propertyReader.VerifyDataProperties(QdrantVectorStoreRecordFieldMapping.s_supportedDataTypes, supportEnumerable: true);
+        propertyReader.VerifyVectorProperties(QdrantVectorStoreRecordFieldMapping.s_supportedVectorTypes);
 
         // Assign.
+        this._propertyReader = propertyReader;
         this._hasNamedVectors = hasNamedVectors;
-        this._keyPropertyInfo = propertiesInfo.KeyProperty;
-        this._dataPropertiesInfo = propertiesInfo.DataProperties;
-        this._vectorPropertiesInfo = propertiesInfo.VectorProperties;
-        this._storagePropertyNames = storagePropertyNames;
-
-        // Get json storage names and store for later use.
-        this._jsonPropertyNames = VectorStoreRecordPropertyReader.BuildPropertyNameToJsonPropertyNameMap(propertiesInfo, typeof(TRecord), JsonSerializerOptions.Default);
     }
 
     /// <inheritdoc />
     public PointStruct MapFromDataToStorageModel(TRecord dataModel)
     {
         PointId pointId;
-        if (this._keyPropertyInfo.PropertyType == typeof(ulong))
+        var keyPropertyInfo = this._propertyReader.KeyPropertyInfo;
+        if (keyPropertyInfo.PropertyType == typeof(ulong))
         {
-            var key = this._keyPropertyInfo.GetValue(dataModel) as ulong? ?? throw new VectorStoreRecordMappingException($"Missing key property {this._keyPropertyInfo.Name} on provided record of type {typeof(TRecord).FullName}.");
+            var key = keyPropertyInfo.GetValue(dataModel) as ulong? ?? throw new VectorStoreRecordMappingException($"Missing key property {keyPropertyInfo.Name} on provided record of type {typeof(TRecord).FullName}.");
             pointId = new PointId { Num = key };
         }
-        else if (this._keyPropertyInfo.PropertyType == typeof(Guid))
+        else if (keyPropertyInfo.PropertyType == typeof(Guid))
         {
-            var key = this._keyPropertyInfo.GetValue(dataModel) as Guid? ?? throw new VectorStoreRecordMappingException($"Missing key property {this._keyPropertyInfo.Name} on provided record of type {typeof(TRecord).FullName}.");
+            var key = keyPropertyInfo.GetValue(dataModel) as Guid? ?? throw new VectorStoreRecordMappingException($"Missing key property {keyPropertyInfo.Name} on provided record of type {typeof(TRecord).FullName}.");
             pointId = new PointId { Uuid = key.ToString("D") };
         }
         else
         {
-            throw new VectorStoreRecordMappingException($"Unsupported key type {this._keyPropertyInfo.PropertyType.FullName} for key property {this._keyPropertyInfo.Name} on provided record of type {typeof(TRecord).FullName}.");
+            throw new VectorStoreRecordMappingException($"Unsupported key type {keyPropertyInfo.PropertyType.FullName} for key property {keyPropertyInfo.Name} on provided record of type {typeof(TRecord).FullName}.");
         }
 
         // Create point.
@@ -97,9 +72,9 @@ internal sealed class QdrantVectorStoreRecordMapper<TRecord> : IVectorStoreRecor
         };
 
         // Add point payload.
-        foreach (var dataPropertyInfo in this._dataPropertiesInfo)
+        foreach (var dataPropertyInfo in this._propertyReader.DataPropertiesInfo)
         {
-            var propertyName = this._storagePropertyNames[dataPropertyInfo.Name];
+            var propertyName = this._propertyReader.GetStoragePropertyName(dataPropertyInfo.Name);
             var propertyValue = dataPropertyInfo.GetValue(dataModel);
             pointStruct.Payload.Add(propertyName, QdrantVectorStoreRecordFieldMapping.ConvertToGrpcFieldValue(propertyValue));
         }
@@ -108,9 +83,9 @@ internal sealed class QdrantVectorStoreRecordMapper<TRecord> : IVectorStoreRecor
         if (this._hasNamedVectors)
         {
             var namedVectors = new NamedVectors();
-            foreach (var vectorPropertyInfo in this._vectorPropertiesInfo)
+            foreach (var vectorPropertyInfo in this._propertyReader.VectorPropertiesInfo)
             {
-                var propertyName = this._storagePropertyNames[vectorPropertyInfo.Name];
+                var propertyName = this._propertyReader.GetStoragePropertyName(vectorPropertyInfo.Name);
                 var propertyValue = vectorPropertyInfo.GetValue(dataModel);
                 if (propertyValue is not null)
                 {
@@ -124,7 +99,7 @@ internal sealed class QdrantVectorStoreRecordMapper<TRecord> : IVectorStoreRecor
         else
         {
             // We already verified in the constructor via FindProperties that there is exactly one vector property when not using named vectors.
-            var vectorPropertyInfo = this._vectorPropertiesInfo.First();
+            var vectorPropertyInfo = this._propertyReader.FirstVectorPropertyInfo!;
             if (vectorPropertyInfo.GetValue(dataModel) is ReadOnlyMemory<float> floatROM)
             {
                 pointStruct.Vectors.Vector = floatROM.ToArray();
@@ -145,11 +120,11 @@ internal sealed class QdrantVectorStoreRecordMapper<TRecord> : IVectorStoreRecor
         var keyPropertyValue = storageModel.Id.HasNum ? storageModel.Id.Num as object : new Guid(storageModel.Id.Uuid) as object;
 
         // Construct the output record.
-        var outputRecord = (TRecord)this._constructorInfo.Invoke(null);
+        var outputRecord = (TRecord)this._propertyReader.ParameterLessConstructorInfo.Invoke(null);
 
         // Set Key.
         var keyPropertyInfoWithValue = new KeyValuePair<PropertyInfo, object?>(
-                this._keyPropertyInfo,
+                this._propertyReader.KeyPropertyInfo,
                 keyPropertyValue);
         VectorStoreRecordMapping.SetPropertiesOnRecord(
             outputRecord,
@@ -161,8 +136,8 @@ internal sealed class QdrantVectorStoreRecordMapper<TRecord> : IVectorStoreRecor
             if (this._hasNamedVectors)
             {
                 var propertiesInfoWithValues = VectorStoreRecordMapping.BuildPropertiesInfoWithValues(
-                    this._vectorPropertiesInfo,
-                    this._storagePropertyNames,
+                    this._propertyReader.VectorPropertiesInfo,
+                    this._propertyReader.StoragePropertyNamesMap,
                     storageModel.Vectors.Vectors_.Vectors,
                     (Vector vector, Type targetType) => new ReadOnlyMemory<float>(vector.Data.ToArray()));
                 VectorStoreRecordMapping.SetPropertiesOnRecord(outputRecord, propertiesInfoWithValues);
@@ -170,7 +145,7 @@ internal sealed class QdrantVectorStoreRecordMapper<TRecord> : IVectorStoreRecor
             else
             {
                 var propertyInfoWithValue = new KeyValuePair<PropertyInfo, object?>(
-                        this._vectorPropertiesInfo[0],
+                        this._propertyReader.FirstVectorPropertyInfo!,
                         new ReadOnlyMemory<float>(storageModel.Vectors.Vector.Data.ToArray()));
                 VectorStoreRecordMapping.SetPropertiesOnRecord(
                     outputRecord,
@@ -179,16 +154,13 @@ internal sealed class QdrantVectorStoreRecordMapper<TRecord> : IVectorStoreRecor
         }
 
         // Set each data property.
-        foreach (var dataProperty in this._dataPropertiesInfo)
-        {
-            var propertiesInfoWithValues = VectorStoreRecordMapping.BuildPropertiesInfoWithValues(
-                this._dataPropertiesInfo,
-                this._storagePropertyNames,
-                storageModel.Payload,
-                (Value grpcValue, Type targetType) =>
-                    QdrantVectorStoreRecordFieldMapping.ConvertFromGrpcFieldValueToNativeType(grpcValue, targetType));
-            VectorStoreRecordMapping.SetPropertiesOnRecord(outputRecord, propertiesInfoWithValues);
-        }
+        var dataPropertiesInfoWithValues = VectorStoreRecordMapping.BuildPropertiesInfoWithValues(
+            this._propertyReader.DataPropertiesInfo,
+            this._propertyReader.StoragePropertyNamesMap,
+            storageModel.Payload,
+            (Value grpcValue, Type targetType) =>
+                QdrantVectorStoreRecordFieldMapping.ConvertFromGrpcFieldValueToNativeType(grpcValue, targetType));
+        VectorStoreRecordMapping.SetPropertiesOnRecord(outputRecord, dataPropertiesInfoWithValues);
 
         return outputRecord;
     }
