+++ dotnet/src/Connectors/Connectors.Qdrant.UnitTests/QdrantGenericDataModelMapperTests.cs
@@ -69,7 +69,8 @@ public class QdrantGenericDataModelMapperTests
     public void MapFromDataToStorageModelMapsAllSupportedTypes(bool hasNamedVectors)
     {
         // Arrange.
-        var sut = new QdrantGenericDataModelMapper(hasNamedVectors ? s_multiVectorStoreRecordDefinition : s_singleVectorStoreRecordDefinition, hasNamedVectors);
+        var reader = new VectorStoreRecordPropertyReader(typeof(VectorStoreGenericDataModel<ulong>), hasNamedVectors ? s_multiVectorStoreRecordDefinition : s_singleVectorStoreRecordDefinition, null);
+        var sut = new QdrantGenericDataModelMapper(reader, hasNamedVectors);
         var dataModel = new VectorStoreGenericDataModel<ulong>(1ul)
         {
             Data =
@@ -159,7 +160,8 @@ public class QdrantGenericDataModelMapperTests
             },
         };
 
-        var sut = (IVectorStoreRecordMapper<VectorStoreGenericDataModel<Guid>, PointStruct>)new QdrantGenericDataModelMapper(vectorStoreRecordDefinition, hasNamedVectors);
+        var reader = new VectorStoreRecordPropertyReader(typeof(VectorStoreGenericDataModel<Guid>), vectorStoreRecordDefinition, null);
+        var sut = (IVectorStoreRecordMapper<VectorStoreGenericDataModel<Guid>, PointStruct>)new QdrantGenericDataModelMapper(reader, hasNamedVectors);
 
         // Act
         var storageModel = sut.MapFromDataToStorageModel(dataModel);
@@ -177,7 +179,8 @@ public class QdrantGenericDataModelMapperTests
     public void MapFromStorageToDataModelMapsAllSupportedTypes(bool hasNamedVectors)
     {
         // Arrange
-        var sut = new QdrantGenericDataModelMapper(hasNamedVectors ? s_multiVectorStoreRecordDefinition : s_singleVectorStoreRecordDefinition, hasNamedVectors);
+        var reader = new VectorStoreRecordPropertyReader(typeof(VectorStoreGenericDataModel<ulong>), hasNamedVectors ? s_multiVectorStoreRecordDefinition : s_singleVectorStoreRecordDefinition, null);
+        var sut = new QdrantGenericDataModelMapper(reader, hasNamedVectors);
         var storageModel = new PointStruct()
         {
             Id = new PointId() { Num = 1 },
@@ -289,7 +292,8 @@ public class QdrantGenericDataModelMapperTests
             storageModel.Vectors.Vector = new Vector() { Data = { 1.0f, 2.0f, 3.0f } };
         }
 
-        var sut = (IVectorStoreRecordMapper<VectorStoreGenericDataModel<Guid>, PointStruct>)new QdrantGenericDataModelMapper(vectorStoreRecordDefinition, hasNamedVectors);
+        var reader = new VectorStoreRecordPropertyReader(typeof(VectorStoreGenericDataModel<Guid>), vectorStoreRecordDefinition, null);
+        var sut = (IVectorStoreRecordMapper<VectorStoreGenericDataModel<Guid>, PointStruct>)new QdrantGenericDataModelMapper(reader, hasNamedVectors);
 
         // Act
         var dataModel = sut.MapFromStorageToDataModel(storageModel, new StorageToDataModelMapperOptions());
@@ -315,7 +319,8 @@ public class QdrantGenericDataModelMapperTests
             },
         };
 
-        var sut = new QdrantGenericDataModelMapper(vectorStoreRecordDefinition, false);
+        var reader = new VectorStoreRecordPropertyReader(typeof(VectorStoreGenericDataModel<ulong>), vectorStoreRecordDefinition, null);
+        var sut = new QdrantGenericDataModelMapper(reader, false);
 
         var dataModel = new VectorStoreGenericDataModel<ulong>(1ul)
         {
@@ -346,7 +351,8 @@ public class QdrantGenericDataModelMapperTests
             },
         };
 
-        var sut = new QdrantGenericDataModelMapper(vectorStoreRecordDefinition, false);
+        var reader = new VectorStoreRecordPropertyReader(typeof(VectorStoreGenericDataModel<ulong>), vectorStoreRecordDefinition, null);
+        var sut = new QdrantGenericDataModelMapper(reader, false);
 
         var dataModel = new VectorStoreGenericDataModel<ulong>(1ul)
         {
@@ -376,7 +382,8 @@ public class QdrantGenericDataModelMapperTests
             },
         };
 
-        var sut = new QdrantGenericDataModelMapper(vectorStoreRecordDefinition, false);
+        var reader = new VectorStoreRecordPropertyReader(typeof(VectorStoreGenericDataModel<ulong>), vectorStoreRecordDefinition, null);
+        var sut = new QdrantGenericDataModelMapper(reader, false);
 
         var storageModel = new PointStruct()
         {
